+++ permission/src/commonMain/kotlin/tech/kotlinlang/permission/PermissionHelper.kt
@@ -3,4 +3,5 @@ package tech.kotlinlang.permission
 interface PermissionHelper {
     suspend fun <T> checkIsPermissionGranted(permission: Permission<T>): T
     suspend fun <T> requestForPermission(permission: Permission<T>): T
+    fun openSettings()
 }